+++ hermes-management/src/test/groovy/pl/allegro/tech/hermes/management/infrastructure/zookeeper/ZookeeperClientManagerTest.groovy
@@ -105,8 +105,4 @@ class ZookeeperClientManagerTest extends MultiZookeeperIntegrationTest {
         clusterProperties.setDatacenter(dcName)
         return clusterProperties
     }
-
-    static findClientByDc(List<ZookeeperClient> clients, String dcName) {
-        return clients.find { it.datacenterName == dcName }
-    }
 }