+++ hermes-test-helper/src/main/java/pl/allegro/tech/hermes/test/helper/endpoint/Waiter.java
@@ -23,20 +23,31 @@ public class Waiter {
 
     public void untilSubscriptionCreated(Topic topic, String subscription, boolean isTracked) {
         waitAtMost(adjust(Duration.ONE_MINUTE)).until(() ->
-            endpoints.subscription().list(topic.getQualifiedName(), isTracked).contains(subscription)
+                endpoints.subscription().list(topic.getQualifiedName(), isTracked).contains(subscription)
+        );
+    }
+
+    public void untilSubscriptionRemoved(Subscription sub) {
+        waitAtMost(adjust(Duration.ONE_MINUTE)).until(() ->
+                !endpoints.subscription().list(sub.getQualifiedTopicName(), sub.isTrackingEnabled()).contains(sub.getName())
         );
     }
 
     public void untilGroupCreated(String group) {
         waitAtMost(adjust(Duration.ONE_MINUTE)).until(() ->
-            endpoints.group().list().contains(group)
+                endpoints.group().list().contains(group)
         );
     }
 
     public void untilTopicCreated(Topic topic) {
         waitAtMost(adjust(Duration.ONE_MINUTE)).until(() ->
-            endpoints.findTopics(topic, topic.isTrackingEnabled()).contains(topic.getQualifiedName())
+                endpoints.findTopics(topic, topic.isTrackingEnabled()).contains(topic.getQualifiedName())
         );
+    }
+
+    public void untilTopicRemoved(Topic topic) {
+        waitAtMost(adjust(Duration.ONE_MINUTE)).until(() ->
+                !endpoints.findTopics(topic, topic.isTrackingEnabled()).contains(topic.getQualifiedName()));
     }
 
     public void untilSchemaCreated(Topic topic) {