+++ integration/src/integration/java/pl/allegro/tech/hermes/integration/KafkaRetransmissionServiceTest.java
@@ -2,8 +2,6 @@ package pl.allegro.tech.hermes.integration;
 
 import com.google.common.collect.ImmutableListMultimap;
 import com.google.common.collect.Multimaps;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 import pl.allegro.tech.hermes.api.ContentType;
@@ -11,12 +9,8 @@ import pl.allegro.tech.hermes.api.PatchData;
 import pl.allegro.tech.hermes.api.Subscription;
 import pl.allegro.tech.hermes.api.Topic;
 import pl.allegro.tech.hermes.common.kafka.offset.PartitionOffset;
-import pl.allegro.tech.hermes.integration.env.ManagementStarter;
-import pl.allegro.tech.hermes.integration.helper.Waiter;
 import pl.allegro.tech.hermes.management.infrastructure.kafka.MultiDCOffsetChangeSummary;
 import pl.allegro.tech.hermes.test.helper.avro.AvroUser;
-import pl.allegro.tech.hermes.test.helper.endpoint.HermesAPIOperations;
-import pl.allegro.tech.hermes.test.helper.endpoint.HermesEndpoints;
 import pl.allegro.tech.hermes.test.helper.endpoint.RemoteServiceEndpoint;
 import pl.allegro.tech.hermes.test.helper.message.TestMessage;
 
@@ -41,27 +35,9 @@ public class KafkaRetransmissionServiceTest extends IntegrationTest {
     private final AvroUser user = new AvroUser();
     private Clock clock = Clock.systemDefaultZone();
 
-    private static String RETRANSMISSION_ENV = "retransmission";
-    private static int MANAGEMENT_PORT = 18083;
-    private static String MANAGEMENT_ENDPOINT_URL = "http://localhost:" + MANAGEMENT_PORT + "/";
-
-    private ManagementStarter managementStarter;
-    private HermesEndpoints management;
-    private HermesAPIOperations operations;
-    private Waiter wait;
-
     @BeforeMethod
     public void initializeAlways() {
         remoteService = new RemoteServiceEndpoint(services().serviceMock());
-    }
-
-    @BeforeClass
-    void beforeClass() throws Exception {
-        managementStarter = new ManagementStarter(MANAGEMENT_PORT, RETRANSMISSION_ENV);
-        managementStarter.start();
-        management = new HermesEndpoints(MANAGEMENT_ENDPOINT_URL, CONSUMER_ENDPOINT_URL);
-        wait = new Waiter(management, services().zookeeper(), brokerOperations, PRIMARY_KAFKA_CLUSTER_NAME, KAFKA_NAMESPACE);
-        operations = new HermesAPIOperations(management, wait);
     }
 
     @Test
@@ -161,11 +137,6 @@ public class KafkaRetransmissionServiceTest extends IntegrationTest {
 
         assertThat(offsets.jsonPartitionOffsets.stream().collect(summingLong(PartitionOffset::getOffset))).isEqualTo(1);
         assertThat(offsets.avroPartitionOffsets.stream().collect(summingLong(PartitionOffset::getOffset))).isEqualTo(0);
-    }
-
-    @AfterClass
-    public void afterClass() throws Exception {
-        managementStarter.stop();
     }
 
     private void sendAvroMessageOnTopic(Topic topic, TestMessage message) {