+++ CHANGELOG.md
@@ -2,9 +2,29 @@
 
 ### ...
 
+## 1.2.0 (24.07.2019)
+
+### Enhancements
+
+#### ([1060](https://github.com/allegro/hermes/pull/1060)) Added support for JDK 11
+
+#### ([1071](https://github.com/allegro/hermes/pull/1071)) Counter values summed from all zookeeper clusters
+
+Counters published, delivered, discarded and volume are kept in zookeeper. Thanks to this their values aren't reset after Hermes restart. 
+However, recently multi-zookeeper feature was introduced in Hermes allowing to have independent Hermes Zookeeper clusters per DC. 
+Because of this when a client sends a request about some topic or subscription metrics to hermes-management then it receives 
+metrics summed from all zookeeper clusters.
+
+#### ([1069](https://github.com/allegro/hermes/pull/1069)) Added storage health check metrics
+
+Since we have [#1052](https://github.com/allegro/hermes/issues/1052), the next step was to introduce metrics 
+for storage (zookeeper clusters) health checks. They are added to MeterRegistry as `storage-health-check.successful` 
+and `storage-health-check.failed` counters.
+
 ## 1.1.2 (08.07.2019)
 
 ### Enhancements
+
 #### ([1061](https://github.com/allegro/hermes/pull/1061)) Introduced HdrHistogram-based metrics reservoir
 
 Introduced [HdrHistogram](http://hdrhistogram.org/)–based implementation of reservoir in metrics. HdrHistogram-based